+++ lib/src/astar_euclidean.dart
@@ -1,6 +1,6 @@
 import 'dart:math';
 
-import '../astar_dart.dart';
+import 'package:astar_dart/astar_dart.dart';
 
 class AStarEuclidean extends AstarGrid {
   AStarEuclidean({
@@ -11,13 +11,10 @@ class AStarEuclidean extends AstarGrid {
 
   @override
   List<ANode> findPath({
-    void Function(List<Point<int>>)? doneList,
+    void Function(List<Point<int>>)? visited,
     required ({int x, int y}) start,
     required ({int x, int y}) end,
   }) {
-    this.doneList.clear();
-    waitList.clear();
-
     if (grid[end.x][end.y].barrier.isBlock) {
       return [];
     }
@@ -28,7 +25,7 @@ class AStarEuclidean extends AstarGrid {
     if (_isNeighbors(start, end)) {
       return [];
     }
-    ANode? winner = _getWinner(
+    ANode? winner = getWinner(
       startNode,
       endNode,
     );
@@ -44,7 +41,10 @@ class AStarEuclidean extends AstarGrid {
         nodeAux = nodeAux.parent!;
       }
     }
-    doneList?.call(this.doneList.map((e) => Point(e.x, e.y)).toList());
+    visited?.call(doneList.map((e) => Point(e.x, e.y)).toList());
+
+    doneList.clear();
+    waitList.clear();
 
     if (winner == null && !_isNeighbors(start, end)) {
       path.clear();
@@ -54,45 +54,12 @@ class AStarEuclidean extends AstarGrid {
   }
 
 //----------------------------------------------------------------------
-  ANode? _getWinner(ANode current, ANode end) {
-    if (end == current) return current;
-    // first iteration
-    for (var n in current.neighbors) {
-      if (n.parent == null) {
-        _checkDistance(n, end, parent: current);
-      }
-      if (!doneList.contains(n)) {
-        waitList.add(n);
-        doneList.add(n);
-      }
-    }
-    waitList.sort((a, b) => b.compareTo(a));
 
-    // loop iteration
-    while (waitList.isNotEmpty) {
-      final c = waitList.removeLast();
-      if (end == c) return c;
-      for (var n in c.neighbors) {
-        if (n.parent == null) {
-          _checkDistance(n, end, parent: c);
-        }
-        if (!doneList.contains(n)) {
-          waitList.add(n);
-          doneList.add(n);
-        }
-      }
-      waitList.sort((a, b) => b.compareTo(a));
-    }
-
-    return null;
-  }
-
-//----------------------------------------------------------------------
-
-  void _checkDistance(ANode current, ANode end, {required ANode parent}) {
+  @override
+  void analyzeDistance(ANode current, ANode end, {required ANode parent}) {
     current.parent = parent;
     current.g = parent.g + current.weight;
-    current.h = _distance(current, end) * 1.0;
+    current.h = _distance(current, end) * 1.5;
   }
 
   int _distance(ANode a, ANode b) {
@@ -115,9 +82,10 @@ class AStarEuclidean extends AstarGrid {
     final maxY = grid.first.length - 1;
     for (var row in grid.array) {
       for (ANode node in row) {
+        // node.reset();
         node.parent = null;
-        node.g = 0.0;
         node.h = 0.0;
+        node.g = 0.0;
         node.neighbors.clear();
         _chainNeighbors(node, maxX: maxX, maxY: maxY);
       }
@@ -130,14 +98,14 @@ class AStarEuclidean extends AstarGrid {
     if (y > 0) {
       // Top
       final neighbor = grid[x][y - 1];
-      if (!grid[x][y - 1].barrier.isBlock) {
+      if (neighbor.barrier != Barrier.block) {
         node.neighbors.add(neighbor);
       }
     }
     if (y < maxY) {
       // Bottom
       final neighbor = grid[x][y + 1];
-      if (!grid[x][y + 1].barrier.isBlock) {
+      if (neighbor.barrier != Barrier.block) {
         node.neighbors.add(neighbor);
       }
     }
@@ -145,14 +113,14 @@ class AStarEuclidean extends AstarGrid {
     if (x > 0) {
       // Left
       final neighbor = grid[x - 1][y];
-      if (!grid[x - 1][y].barrier.isBlock) {
+      if (neighbor.barrier != Barrier.block) {
         node.neighbors.add(neighbor);
       }
     }
     if (x < maxX) {
       // Right
       final neighbor = grid[x + 1][y];
-      if (!grid[x + 1][y].barrier.isBlock) {
+      if (neighbor.barrier != Barrier.block) {
         node.neighbors.add(neighbor);
       }
     }
@@ -161,21 +129,21 @@ class AStarEuclidean extends AstarGrid {
     if (x > 0 && y > 0) {
       // Top-Left
       final neighbor = grid[x - 1][y - 1];
-      if (!grid[x - 1][y - 1].barrier.isBlock) {
+      if (neighbor.barrier != Barrier.block) {
         node.neighbors.add(neighbor);
       }
     }
     if (x > 0 && y < maxY) {
       // Bottom-Left
       final neighbor = grid[x - 1][y + 1];
-      if (!grid[x - 1][y + 1].barrier.isBlock) {
+      if (neighbor.barrier != Barrier.block) {
         node.neighbors.add(neighbor);
       }
     }
     if (x < maxX && y > 0) {
       // Top-Right
       final neighbor = grid[x + 1][y - 1];
-      if (!grid[x + 1][y - 1].barrier.isBlock) {
+      if (neighbor.barrier != Barrier.block) {
         node.neighbors.add(neighbor);
       }
     }
@@ -183,7 +151,7 @@ class AStarEuclidean extends AstarGrid {
     if (x < maxX && y < maxY) {
       // Bottom-Right
       final neighbor = grid[x + 1][y + 1];
-      if (!grid[x + 1][y + 1].barrier.isBlock) {
+      if (neighbor.barrier != Barrier.block) {
         node.neighbors.add(neighbor);
       }
     }